+++ frontend/src/pages/MainPage.tsx
@@ -1,20 +1,17 @@
-import { Link } from 'react-router-dom';
 import { styled } from 'styled-components';
 import CarouselItem from '@/features/songs/components/CarouselItem';
 import CollectionCarousel from '@/features/songs/components/CollectionCarousel';
-import SongItem from '@/features/songs/components/SongItem';
+import SongItemList from '@/features/songs/components/SongItemList';
 import Spacing from '@/shared/components/Spacing';
 import SRHeading from '@/shared/components/SRHeading';
-import ROUTE_PATH from '@/shared/constants/path';
 import useFetch from '@/shared/hooks/useFetch';
 import fetcher from '@/shared/remotes';
-import type { PopularSong, VotingSong } from '@/features/songs/types/Song.type';
+import type { VotingSong } from '@/features/songs/types/Song.type';
 
 const MainPage = () => {
-  const { data: popularSongs } = useFetch<PopularSong[]>(() => fetcher('/songs/high-liked', 'GET'));
   const { data: votingSongs } = useFetch<VotingSong[]>(() => fetcher('/voting-songs', 'GET'));
 
-  if (!popularSongs || !votingSongs) return null;
+  if (!votingSongs) return null;
 
   const isEmptyVotingSongs = votingSongs.length === 0;
 
@@ -35,26 +32,7 @@ const MainPage = () => {
         )}
       </CollectionCarousel>
       <Spacing direction="vertical" size={24} />
-      <Title>킬링파트 좋아요 많은순</Title>
-      <Spacing direction="vertical" size={16} />
-      <PopularSongList>
-        {popularSongs.map(({ id, albumCoverUrl, title, singer, totalLikeCount }, i) => (
-          <Li key={id}>
-            <StyledLink
-              to={`${ROUTE_PATH.SONG_DETAILS}/${id}`}
-              aria-label={`킬링파트 투표순 ${i + 1}등 ${singer} ${title}`}
-            >
-              <SongItem
-                rank={i + 1}
-                albumCoverUrl={albumCoverUrl}
-                title={title}
-                singer={singer}
-                totalLikeCount={totalLikeCount}
-              />
-            </StyledLink>
-          </Li>
-        ))}
-      </PopularSongList>
+      <SongItemList />
     </Container>
   );
 };
@@ -76,34 +54,11 @@ const Container = styled.div`
   }
 `;
 
-const Li = styled.li`
-  max-width: 250px;
-`;
-
-const StyledLink = styled(Link)`
-  width: 100%;
-
-  &:hover,
-  &:focus {
-    background-color: ${({ theme }) => theme.color.secondary};
-  }
-`;
-
 const Title = styled.h2`
   align-self: flex-start;
   font-size: 20px;
   font-weight: 700;
   color: white;
-`;
-
-const PopularSongList = styled.ol`
-  overflow-x: scroll;
-  display: flex;
-  flex-direction: row;
-  gap: 8px;
-  align-items: flex-start;
-
-  width: 100%;
 `;
 
 const EmptyMessage = styled.li`
